requirements.txt
selenium==4.15.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pyyaml==6.0.1
flask==2.3.3
webdriver-manager==4.0.1
urllib3==1.26.18
